Extract template expressions into own file
