+++ decoder/expr_any_ref_origins.go
@@ -246,40 +246,3 @@ func (a Any) refOriginsForOperatorExpr(ctx context.Context, allowSelfRefs bool) 
 
 	return origins, false
 }
-
-func (a Any) refOriginsForTemplateExpr(ctx context.Context, allowSelfRefs bool) (reference.Origins, bool) {
-	origins := make(reference.Origins, 0)
-
-	switch eType := a.expr.(type) {
-	case *hclsyntax.TemplateExpr:
-		if eType.IsStringLiteral() {
-			return nil, false
-		}
-
-		for _, partExpr := range eType.Parts {
-			cons := schema.AnyExpression{
-				OfType: cty.String,
-			}
-			expr := newExpression(a.pathCtx, partExpr, cons)
-
-			if e, ok := expr.(ReferenceOriginsExpression); ok {
-				origins = append(origins, e.ReferenceOrigins(ctx, allowSelfRefs)...)
-			}
-		}
-
-		return origins, true
-	case *hclsyntax.TemplateWrapExpr:
-		cons := schema.AnyExpression{
-			OfType: cty.String,
-		}
-		expr := newExpression(a.pathCtx, eType.Wrapped, cons)
-
-		if e, ok := expr.(ReferenceOriginsExpression); ok {
-			origins = append(origins, e.ReferenceOrigins(ctx, allowSelfRefs)...)
-		}
-
-		return origins, true
-	}
-
-	return origins, false
-}
